Move encrypted test files to cluster package
